[CI] Add `ShellCheck` workflow to Buildkite
Summary: Relevant Linear issue [[ https://linear.app/comm/issue/ENG-1296/add-shellcheck-or-something-similar-to-the-ci | here ]].

This diff adds [[ https://www.shellcheck.net/ | ShellCheck ]] to the CI, a static analysis tool for shell scripts that provides warnings and errors.

We install `ShellCheck` and also set `globstar` so we can use `**` to recursively match all `.sh` files in the repository. These files are then evaluated by `ShellCheck`.

Test Plan: Ran in Buildkite and worked as expected. Currently the build is failing since we have so many shell scripts that contain problematic code, but that's expected!

Reviewers: atul, palys-swm, def-au1t, karol-bisztyga, varun

Subscribers: