REmove unnecessary code from message reducer. add rehydrate transform to insert messages field to messageStore since it is essential to avoid null/undefined bugs
Summary: This differential add rehydrate transform that inserts {} as messages field in messageStore. Previously this was done when handling rehydrateActionType. Now we want to remove this code, but we still need messageStore to have this field to avoid null/undefined errors.

Test Plan: This differential does not change app functionality, but only the way it works. So we expect the app to work exactly the same after this change.

Reviewers: palys-swm, atul, def-au1t

Subscribers: