Remove unnecessary (dead code) code from message reducer.
Summary: This differential removed code that was used to modify messageStore on rehydrateActionType. Now that we do not persist messages in messageStore in redux, messages are always {} on rehydrateActionType, so the code never get executed.

Test Plan: This differential does not change app functionality, but only the way it works. So we expect the app to work exactly the same after this change.

Reviewers: palys-swm, atul, def-au1t

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4507