[services] Tunnelbroker - Add Cargo manifest file
Summary:

This diff introducing a Rust Cargo manifest file for the notification library wrapper in Rust for using it from C++ using corrosion tool.

Test Plan:

No tests, this is a manifest-only file. This file will be updated with the `dependencies` and build using `cargo build` in a follow-up diffs of the
stack.

Reviewers:

Subscribers: